change mask dtype to uint 8

+++ tensorflow_datasets/image/.ipynb_checkpoints/brats_2015-checkpoint.py
@@ -36,7 +36,7 @@ class Brats2015(tfds.core.GeneratorBasedBuilder):
         # tfds.features.FeatureConnectors
         features=tfds.features.FeaturesDict({
         #The MRI image
-        'image' : tfds.features.Tensor(shape=(240,240),dtype=tf.int16),
+        'image' : tfds.features.Tensor(shape=(240,240),dtype=tf.int8),
         ## The mask
         'mask' : tfds.features.Tensor(shape=(512,512),dtype = tf.int16),
         ## Tumor Type
